Moodle 4.3 upgrade

+++ admin/tool/uploadcourse/classes/course.php
@@ -932,17 +932,22 @@ class tool_uploadcourse_course {
 
         foreach ($enrolmentdata as $method => $options) {
 
-            if (isset($options['role'])) {
-                $role = $options['role'];
+            if (isset($options['role']) || isset($options['roleid'])) {
+                if (isset($options['role'])) {
+                    $role = $options['role'];
+                    $roleid = $DB->get_field('role', 'id', ['shortname' => $role], MUST_EXIST);
+                } else {
+                    $roleid = $options['roleid'];
+                    $role = $DB->get_field('role', 'shortname', ['id' => $roleid], MUST_EXIST);
+                }
                 if ($courseid) {
-                    if (!$this->validate_role_context($courseid, $role)) {
+                    if (!$this->validate_role_context($courseid, $roleid)) {
                         $errors['contextrolenotallowed'] = new lang_string('contextrolenotallowed', 'core_role', $role);
 
                         break;
                     }
                 } else {
                     // We can at least check that context level is correct while actual context not exist.
-                    $roleid = $DB->get_field('role', 'id', ['shortname' => $role], MUST_EXIST);
                     if (!$this->validate_role_context_level($roleid)) {
                         $errors['contextrolenotallowed'] = new lang_string('contextrolenotallowed', 'core_role', $role);
 
@@ -951,9 +956,13 @@ class tool_uploadcourse_course {
                 }
             }
 
-            if ($courseid) {
-                $plugin = $enrolmentplugins[$method];
+            $plugin = $enrolmentplugins[$method];
+            $errors += $plugin->validate_enrol_plugin_data($options, $courseid);
+            if ($errors) {
+                break;
+            }
 
+            if ($courseid) {
                 // Find matching instances by enrolment method.
                 $methodinstances = array_filter($instances, static function (stdClass $instance) use ($method) {
                     return (strcmp($instance->enrol, $method) == 0);
@@ -1035,7 +1044,7 @@ class tool_uploadcourse_course {
                     $plugin = $enrolmentplugins[$instance->enrol];
 
                     // Ensure user is able to delete the instance.
-                    if ($plugin->can_delete_instance($instance)) {
+                    if ($plugin->can_delete_instance($instance) && $plugin->is_csv_upload_supported()) {
                         $plugin->delete_instance($instance);
                     } else {
                         $this->error('errorcannotdeleteenrolment',
@@ -1047,83 +1056,110 @@ class tool_uploadcourse_course {
                 // Create/update enrolment.
                 $plugin = $enrolmentplugins[$enrolmethod];
 
-                $status = ($todisable) ? ENROL_INSTANCE_DISABLED : ENROL_INSTANCE_ENABLED;
+                if ($plugin->is_csv_upload_supported()) {
+                    $status = ($todisable) ? ENROL_INSTANCE_DISABLED : ENROL_INSTANCE_ENABLED;
+                    $method = $plugin->fill_enrol_custom_fields($method, $course->id);
 
-                // Create a new instance if necessary.
-                if (empty($instance) && $plugin->can_add_instance($course->id)) {
-                    $instanceid = $plugin->add_default_instance($course);
-                    $instance = $DB->get_record('enrol', ['id' => $instanceid]);
-                    $instance->roleid = $plugin->get_config('roleid');
-                    // On creation the user can decide the status.
-                    $plugin->update_status($instance, $status);
-                }
+                    // Create a new instance if necessary.
+                    if (empty($instance) && $plugin->can_add_instance($course->id)) {
+                        $error = $plugin->validate_plugin_data_context($method, $course->id);
+                        if ($error) {
+                            $this->error('contextnotallowed', $error);
+                            break;
+                        }
+                        $instanceid = $plugin->add_default_instance($course);
+                        if (!$instanceid) {
+                            // Add instance with provided fields if plugin supports it.
+                            $instanceid = $plugin->add_custom_instance($course, $method);
+                        }
 
-                // Check if the we need to update the instance status.
-                if ($instance && $status != $instance->status) {
-                    if ($plugin->can_hide_show_instance($instance)) {
-                        $plugin->update_status($instance, $status);
-                    } else {
-                        $this->error('errorcannotdisableenrolment',
-                            new lang_string('errorcannotdisableenrolment', 'tool_uploadcourse',
+                        $instance = $DB->get_record('enrol', ['id' => $instanceid]);
+                        if ($instance) {
+                            $instance->roleid = $plugin->get_config('roleid');
+                            // On creation the user can decide the status.
+                            $plugin->update_status($instance, $status);
+                        }
+                    }
+
+                    // Check if the we need to update the instance status.
+                    if ($instance && $status != $instance->status) {
+                        if ($plugin->can_hide_show_instance($instance)) {
+                            $plugin->update_status($instance, $status);
+                        } else {
+                            $this->error('errorcannotdisableenrolment',
+                                new lang_string('errorcannotdisableenrolment', 'tool_uploadcourse',
+                                    $plugin->get_instance_name($instance)));
+                            break;
+                        }
+                    }
+
+                    if (empty($instance) || !$plugin->can_edit_instance($instance)) {
+                        $this->error('errorcannotcreateorupdateenrolment',
+                            new lang_string('errorcannotcreateorupdateenrolment', 'tool_uploadcourse',
                                 $plugin->get_instance_name($instance)));
-                        break;
-                    }
-                }
 
-                if (empty($instance) || !$plugin->can_edit_instance($instance)) {
-                    $this->error('errorcannotcreateorupdateenrolment',
-                        new lang_string('errorcannotcreateorupdateenrolment', 'tool_uploadcourse',
-                            $plugin->get_instance_name($instance)));
-
-                    break;
-                }
-
-                // Now update values.
-                foreach ($method as $k => $v) {
-                    $instance->{$k} = $v;
-                }
-
-                // Sort out the start, end and date.
-                $instance->enrolstartdate = (isset($method['startdate']) ? strtotime($method['startdate']) : 0);
-                $instance->enrolenddate = (isset($method['enddate']) ? strtotime($method['enddate']) : 0);
-
-                // Is the enrolment period set?
-                if (isset($method['enrolperiod']) && ! empty($method['enrolperiod'])) {
-                    if (preg_match('/^\d+$/', $method['enrolperiod'])) {
-                        $method['enrolperiod'] = (int) $method['enrolperiod'];
-                    } else {
-                        // Try and convert period to seconds.
-                        $method['enrolperiod'] = strtotime('1970-01-01 GMT + ' . $method['enrolperiod']);
-                    }
-                    $instance->enrolperiod = $method['enrolperiod'];
-                }
-                if ($instance->enrolstartdate > 0 && isset($method['enrolperiod'])) {
-                    $instance->enrolenddate = $instance->enrolstartdate + $method['enrolperiod'];
-                }
-                if ($instance->enrolenddate > 0) {
-                    $instance->enrolperiod = $instance->enrolenddate - $instance->enrolstartdate;
-                }
-                if ($instance->enrolenddate < $instance->enrolstartdate) {
-                    $instance->enrolenddate = $instance->enrolstartdate;
-                }
-
-                // Sort out the given role.
-                if (isset($method['role'])) {
-                    $role = $method['role'];
-                    if (!$this->validate_role_context($course->id, $role)) {
-                        $this->error('contextrolenotallowed',
-                            new lang_string('contextrolenotallowed', 'core_role', $role));
                         break;
                     }
 
-                    $roleids = tool_uploadcourse_helper::get_role_ids();
-                    if (isset($roleids[$method['role']])) {
-                        $instance->roleid = $roleids[$method['role']];
-                    }
-                }
+                    // Now update values.
+                    $modifiedinstance = $instance;
 
-                $instance->timemodified = time();
-                $DB->update_record('enrol', $instance);
+                    // Sort out the start, end and date.
+                    $modifiedinstance->enrolstartdate = (isset($method['startdate']) ? strtotime($method['startdate']) : 0);
+                    $modifiedinstance->enrolenddate = (isset($method['enddate']) ? strtotime($method['enddate']) : 0);
+
+                    // Is the enrolment period set?
+                    if (isset($method['enrolperiod']) && !empty($method['enrolperiod'])) {
+                        if (preg_match('/^\d+$/', $method['enrolperiod'])) {
+                            $method['enrolperiod'] = (int)$method['enrolperiod'];
+                        } else {
+                            // Try and convert period to seconds.
+                            $method['enrolperiod'] = strtotime('1970-01-01 GMT + ' . $method['enrolperiod']);
+                        }
+                        $modifiedinstance->enrolperiod = $method['enrolperiod'];
+                    }
+                    if ($instance->enrolstartdate > 0 && isset($method['enrolperiod'])) {
+                        $modifiedinstance->enrolenddate = $instance->enrolstartdate + $method['enrolperiod'];
+                    }
+                    if ($instance->enrolenddate > 0) {
+                        $modifiedinstance->enrolperiod = $instance->enrolenddate - $instance->enrolstartdate;
+                    }
+                    if ($instance->enrolenddate < $instance->enrolstartdate) {
+                        $modifiedinstance->enrolenddate = $instance->enrolstartdate;
+                    }
+
+                    // Sort out the given role.
+                    if (isset($method['role']) || isset($method['roleid'])) {
+                        if (isset($method['role'])) {
+                            $role = $method['role'];
+                            $roleid = $DB->get_field('role', 'id', ['shortname' => $role], MUST_EXIST);
+                        } else {
+                            $roleid = $method['roleid'];
+                            $role = $DB->get_field('role', 'shortname', ['id' => $roleid], MUST_EXIST);
+                        }
+                        if (!$this->validate_role_context($course->id, $roleid)) {
+                            $this->error('contextrolenotallowed',
+                                new lang_string('contextrolenotallowed', 'core_role', $role));
+                            break;
+                        }
+
+                        $roleids = tool_uploadcourse_helper::get_role_ids();
+                        if (in_array($roleid, $roleids)) {
+                            $modifiedinstance->roleid = $roleid;
+                        }
+                    }
+
+                    // Sort out custom instance name.
+                    if (isset($method['name'])) {
+                        $modifiedinstance->name = $method['name'];
+                    }
+
+                    $plugin->update_instance($instance, $modifiedinstance);
+                } else {
+                    $this->error('errorunsupportedmethod',
+                        new lang_string('errorunsupportedmethod', 'tool_uploadcourse',
+                            $enrolmethod));
+                }
             }
         }
     }
@@ -1132,15 +1168,15 @@ class tool_uploadcourse_course {
      * Check if role is allowed in course context
      *
      * @param int $courseid course context.
-     * @param string $role Role.
+     * @param int $roleid Role ID.
      * @return bool
      */
-    protected function validate_role_context(int $courseid, string $role) : bool {
+    protected function validate_role_context(int $courseid, int $roleid) : bool {
         if (empty($this->assignableroles[$courseid])) {
             $coursecontext = \context_course::instance($courseid);
             $this->assignableroles[$courseid] = get_assignable_roles($coursecontext, ROLENAME_SHORT);
         }
-        if (!in_array($role, $this->assignableroles[$courseid])) {
+        if (!array_key_exists($roleid, $this->assignableroles[$courseid])) {
             return false;
         }
         return true;
